+++ nolzo-core/src/main/java/com/noljo/nolzo/notification/application/port/in/HandleNotificationBatchUseCase.java
@@ -1,0 +1,8 @@
+package com.noljo.nolzo.notification.application.port.in;
+
+import com.noljo.nolzo.notification.domain.event.NotificationBatchRequest;
+
+public interface HandleNotificationBatchUseCase {
+
+    void handle(NotificationBatchRequest request);
+}
