+++ nolzo-core/src/main/java/com/noljo/nolzo/notification/domain/event/NotificationBatchRequest.java
@@ -1,0 +1,14 @@
+package com.noljo.nolzo.notification.domain.event;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+public record NotificationBatchRequest(
+        Long eventId,
+        Long eventScheduleId,
+        Long seatId,
+        String seatGrade,
+        LocalDateTime availableAt,
+        List<Long> subscriptionIds
+) {
+}
